Fix some obvious errors

+++ roadrunner/sync_worker_test.go
@@ -181,17 +181,17 @@ func Test_NumExecs(t *testing.T) {
 	if err != nil {
 		t.Errorf("fail to execute payload: error %v", err)
 	}
-	assert.Equal(t, int64(1), w.State().NumExecs())
+	assert.Equal(t, 1, w.State().NumExecs())
 
 	_, err = w.Exec(&Payload{Body: []byte("hello")})
 	if err != nil {
 		t.Errorf("fail to execute payload: error %v", err)
 	}
-	assert.Equal(t, int64(2), w.State().NumExecs())
+	assert.Equal(t, 2, w.State().NumExecs())
 
 	_, err = w.Exec(&Payload{Body: []byte("hello")})
 	if err != nil {
 		t.Errorf("fail to execute payload: error %v", err)
 	}
-	assert.Equal(t, int64(3), w.State().NumExecs())
+	assert.Equal(t, 3, w.State().NumExecs())
 }
